videoflix-backend/requirements.txt
# Videoflix Backend – Python dependencies
# Install: pip install -r requirements.txt

# Django
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
python-dotenv==1.2.1
whitenoise==6.11.0

# Django REST Framework & API
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
uritemplate==4.2.0
PyYAML==6.0.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# Auth (JWT)
djangorestframework-simplejwt==5.5.1
PyJWT==2.11.0

# Database
psycopg2-binary==2.9.11

# Cache & Task Queue (Redis, RQ)
redis==7.1.1
django-redis==6.0.0
django-rq==3.2.2
rq==2.6.1
croniter==6.0.0

# Media & Images
pillow==12.1.1

# Server (Production)
gunicorn==25.0.3

# Development & Debug
django-debug-toolbar==6.2.0

# Utils
asgiref==3.11.1
attrs==25.4.0
click==8.3.1
colorama==0.4.6
packaging==26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
